Getting a Money object with decimal places from a MoneyField
Right now when getting a MoneyField it returns a Money object without the MoneyField decimal places, producing unexpected behaviour when printing/converting to a string. I'm not sure if that works as I don't have the time to test so consider this as an issue with a probable pull request. It's my first pull request so sorry about anything, and have a nice day :)